Fix hash calculation of truncated RLP header
The old implementation was assuming that the nonce and mixHash header fields were the last two fields.
Therefore it was easy to slice off the last 42 bytes and adjust the RLP encoded payload length.

The `baseFeePerGas` field got appended to the existing header structure due to EIP-1559, breaking the simplified calculation.

Hence it is easier to do the hash calculation on the client side in order to avoid RLP encoding and decoding on-chain.